IntelliTest Wizard fix. (#103)
CLR starts initializing static variables in the order in which they appear on file. Once we moved all public methods to the top to satisfy stylecop AssemblyName and RootNamespace which needed to be initialized before AssertType and CollectionAssertType used them were initialized later. Moving to an on-demand model instead. 

Also just including direct dependecies in the vsix. We do not want to carry around assemblies dropped by VS.