Use different way to get the attribute and mocks in tests
The usage of ReflectHelper is a bit nuanced, the generic
method I used originally uses a static method interallly
which prevents it from being mocked. So in order to write
the tests in the same way as the other tests are written
I use the mock to act like the method is decorated with the
correct attribute, and need to use a different method internally
but at least the generic method is not the sigle outlier in that file.