Change NuGet package to use `None` ItemGroup to copy files to output directory
Without this change, files from this nuget package will be included
automatically when the consuming project is packing.

Example error when packing a project:
> C:\Program Files\dotnet\sdk\3.1.200\Sdks\NuGet.Build.Tasks.Pack\build\NuGet.Build.Tasks.Pack.targets(198,5):
warning NU5100: The assembly 'content\Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.dll' is
not inside the 'lib' folder and hence it won't be added as a reference
when the package is installed into a project. Move it into the 'lib'
folder if it needs to be referenced.

Additionally, add the contents to an ItemGroup so they can be seperated later if needed.