Fix issue cannot load Microsoft.TestPlatform.CoreUtilities
When setuping DeploymentEnabled, MSTest fails with `An exception occurred while invoking executor 'executor://mstestadapter/v2': Could not load file or assembly 'Microsoft.TestPlatform.CoreUtilities, Version=15.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a' or one of its dependencies. The system cannot find the file specified.`. After investigation, it seems related to a chage introduced in 2.2.7 where some logs were added, adding a dependency to this dll which cannot be resolved by the custom assembly resolver.

The special appdomain and resolver are meant to isolate and detect dll coming from the user bin folder to enable automatic copy of these dlls so we cannot easily fix the resolver as it could lead to conflicts or missing dlls copied. Simplest solution is to remove the log steps for this class.