Revert change for timeout test execution
Through the process of simplification and unification of behaviors for the different target frameworks, we dropped usage of a thread to use only Task.Run. This has the unexpected side effect of causing .NET framework tests to no longer be run in STA thread.

We are reverting the change for now and are tracking the implementation of a better solution with https://github.com/microsoft/testfx/issues/1674