Allow IEnumerable of tuple for DynamicData source
To avoid breaking public API, we decided to convert the tuple in
`object[]` to preserve it. The current solution relies on `ITuple`
which is only available in NET471+. Given we currently target net462
for full framework, that means that the feature will only be available
for netcore. I'll weigh the benefit/cost ratio of adding net471 build
 in the package so this feature can be available more widely.