Improve user experience of restoring/building/testing
Add open-vs.cmd, restore.sh and restore.cmd. Fix build.cmd, test.cmd
and test.sh.

We will now always try to install the Windows 10 SDK except if the
-skipInstallWindowsSdk switch is provided. This should reduce the
number of issues linked to this SDK being missing.

Provides a script to allow opening the solution ensuring that the
local SDK is used. It won't prevent users to see issue if they open
the sln directly but this should reduce friction, especially after
the contributing readme is up to date.