Add support for STA thread
When run with MSTest runner, when on Windows and the user specifies
`ExecutionApartmentState` entry in the runsettings and the current
thread is not already in the right state, we now create a
global thread that switches to the requested apartment state.

When run on VSTest, we rely on VSTest setting the right global state.

The test method execution is also fixed for .NET Core to rely on thread
when we need to be on STA thread.